+++ tests/tests/cloudinary_/models/test_file.py
@@ -47,6 +47,9 @@ class TestCloudinaryFile(CloudinaryFileResource):
         storage.resource.delete_file()
         storage.resource.delete()
 
+    def test_get_file_folder(self, storage):
+        assert storage.resource.get_file_folder() == self.resource_location
+
     def test_display_name(self, storage):
         assert storage.resource.display_name == self.resource_name