+++ tests/tests/models/test_image.py
@@ -44,6 +44,9 @@ class TestUploadedImage(TestVersatileImageResource):
         storage.resource.delete_file()
         storage.resource.delete()
 
+    def test_get_file_folder(self, storage):
+        assert storage.resource.get_file_folder() == self.resource_location
+
     def test_as_dict(self, storage):
         assert storage.resource.as_dict() == {
             'id': 1,