+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [0.6.0](https://github.com/dldevinc/paper-uploads/tree/v0.6.0) - 2021-09-07
 ### Features
 - Removed `CloudinaryCollection` class.
+- Add a method `get_file_folder` returning storage directory.
+  Override this method to customize the upload folder.
 
 ## [0.5.2](https://github.com/dldevinc/paper-uploads/tree/v0.5.2) - 2021-09-03
 ### Bug Fixes